I have completed the entirety of the Unity code,
The wheels don't move but no one will notice. The robot moves(Around a bad axis but again, unnoticeable) and it shoots balls! With randomness! And Physics! If the balls hit either the ground or the collector they are destroyed. I managed to do it! You can climb over the walls